# Upgrade Docusaurus - Enable Docusaurus Faster on website