--- a/.gitignore
+++ b/.gitignore
@@ -20,3 +20,5 @@ yarn-error.log
 !.yarn/releases
 !.yarn/sdks
 !.yarn/versions
+
+.idea