--- a/.yarnrc.yml
+++ b/.yarnrc.yml
@@ -5,3 +5,13 @@ enableGlobalCache: false
 npmPublishAccess: public
 
 yarnPath: .yarn/releases/yarn-4.0.1.cjs
+
+nodeLinker: node-modules
+
+packageExtensions:
+  "@docusaurus/bundler@*":
+    dependencies:
+      "@docusaurus/faster": "3.5.2-canary-6122"
+  "@docusaurus/faster@*":
+    dependencies:
+      "tslib": "^2.6.0"