--- a/.yarnrc.yml
+++ b/.yarnrc.yml
@@ -11,7 +11,5 @@ nodeLinker: node-modules
 packageExtensions:
   "@docusaurus/bundler@*":
     dependencies:
-      "@docusaurus/faster": "3.5.2-canary-6122"
-  "@docusaurus/faster@*":
-    dependencies:
-      "tslib": "^2.6.0"
+      "@docusaurus/faster": "*"
+