--- a/.yarnrc.yml
+++ b/.yarnrc.yml
@@ -6,10 +6,20 @@ npmPublishAccess: public
 
 yarnPath: .yarn/releases/yarn-4.0.1.cjs
 
-nodeLinker: node-modules
+# nodeLinker: node-modules
 
 packageExtensions:
   "@docusaurus/bundler@*":
     dependencies:
       "@docusaurus/faster": "*"
+  "@docusaurus/theme-common@*":
+    dependencies:
+      "@docusaurus/plugin-content-docs": "*"
+  "terser-webpack-plugin@*":
+    dependencies:
+      "@swc/core": "*"
+  "css-minimizer-webpack-plugin@*":
+    dependencies:
+      "lightningcss": "*"
+