--- a/website-multi-docs/babel.config.js
+++ b/website-multi-docs/babel.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  presets: [require.resolve('@docusaurus/core/lib/babel/preset')],
-};