--- a/website-multi-docs/package.json
+++ b/website-multi-docs/package.json
@@ -14,11 +14,12 @@
     "write-heading-ids": "docusaurus write-heading-ids"
   },
   "dependencies": {
-    "@docusaurus/core": "^3.0.0",
-    "@docusaurus/plugin-content-docs": "^3.0.0",
-    "@docusaurus/preset-classic": "^3.0.0",
-    "@docusaurus/theme-classic": "^3.0.0",
-    "@docusaurus/theme-common": "^3.0.0",
+    "@docusaurus/core": "3.6.0",
+    "@docusaurus/faster": "3.6.0",
+    "@docusaurus/plugin-content-docs": "3.6.0",
+    "@docusaurus/preset-classic": "3.6.0",
+    "@docusaurus/theme-classic": "3.6.0",
+    "@docusaurus/theme-common": "3.6.0",
     "@easyops-cn/docusaurus-search-local": "^workspace:*",
     "@mdx-js/react": "^3.0.0",
     "clsx": "^1.1.1",
@@ -27,7 +28,7 @@
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "^3.0.0"
+    "@docusaurus/module-type-aliases": "3.6.0"
   },
   "browserslist": {
     "production": [