--- a/website/package.json
+++ b/website/package.json
@@ -12,11 +12,11 @@
     "serve": "docusaurus serve"
   },
   "dependencies": {
-    "@docusaurus/core": "3.5.2-canary-6122",
-    "@docusaurus/faster": "3.5.2-canary-6122",
-    "@docusaurus/preset-classic": "3.5.2-canary-6122",
-    "@docusaurus/theme-classic": "3.5.2-canary-6122",
-    "@docusaurus/theme-common": "3.5.2-canary-6122",
+    "@docusaurus/core": "3.6.0",
+    "@docusaurus/faster": "3.6.0",
+    "@docusaurus/preset-classic": "3.6.0",
+    "@docusaurus/theme-classic": "3.6.0",
+    "@docusaurus/theme-common": "3.6.0",
     "@easyops-cn/docusaurus-search-local": "workspace:*",
     "@mdx-js/react": "^3.0.0",
     "clsx": "^1.1.1",