Docusaurus maintainer here 👋 

This is a POC to ensure this plugin is compatible with [Docusaurus Faster](https://github.com/facebook/docusaurus/issues/10556)

I'm not sure how your setup works exactly, not super familiar with Yarn PnP but got some annoying errors due to it 😅 

I'm able to build with Webpack, but Rspack fails with this error I can't solve when PnP is enabled:

> [ERROR] Client bundle compiled with errors therefore further build is impossible.
× Module not found: Can't resolve '/Users/sebastienlorber/Desktop/projects/docusaurus-search-local-easyops-cn/.yarn/__virtual__/@docusaurus-core-virtual-aaf120fccf/0/cache/@docusaurus-core-npm-3.5.2-canary-6122-e777a73518-b778e9de54.zip/node_modules/@docusaurus/core/lib/client/clientEntry.js' in '/Users/sebastienlorber/Desktop/projects/docusaurus-search-local-easyops-cn/website'

And idea?

## Benchmark

Using node modules linker

Cold builds - much faster

```bash
hyperfine --prepare 'yarn workspace website clear' --runs 3 'DOCUSAURUS_FASTER=false yarn build:website' 'DOCUSAURUS_FASTER=true yarn build:website'


Benchmark 1: DOCUSAURUS_FASTER=false yarn build:website
  Time (mean ± σ):     11.864 s ±  0.251 s    [User: 19.999 s, System: 5.471 s]
  Range (min … max):   11.609 s … 12.110 s    3 runs

Benchmark 2: DOCUSAURUS_FASTER=true yarn build:website
  Time (mean ± σ):      3.139 s ±  0.050 s    [User: 4.666 s, System: 0.862 s]
  Range (min … max):    3.106 s …  3.196 s    3 runs

Summary
  DOCUSAURUS_FASTER=true yarn build:website ran
    3.78 ± 0.10 times faster than DOCUSAURUS_FASTER=false yarn build:website
```

Warm builds - same:

```bash
DOCUSAURUS_FASTER=false yarn build:website
hyperfine --runs 3 'DOCUSAURUS_FASTER=false yarn build:website' 'DOCUSAURUS_FASTER=true yarn build:website'


Benchmark 1: DOCUSAURUS_FASTER=false yarn build:website
  Time (mean ± σ):      3.031 s ±  0.066 s    [User: 3.602 s, System: 0.539 s]
  Range (min … max):    2.990 s …  3.107 s    3 runs

  Warning: The first benchmarking run for this command was significantly slower than the rest (3.107 s). This could be caused by (filesystem) caches that were not filled until after the first run. You should consider using the '--warmup' option to fill those caches before the actual benchmark. Alternatively, use the '--prepare' option to clear the caches before each timing run.

Benchmark 2: DOCUSAURUS_FASTER=true yarn build:website
  Time (mean ± σ):      3.120 s ±  0.084 s    [User: 4.671 s, System: 0.859 s]
  Range (min … max):    3.065 s …  3.217 s    3 runs

Summary
  DOCUSAURUS_FASTER=false yarn build:website ran
    1.03 ± 0.04 times faster than DOCUSAURUS_FASTER=true yarn build:website
```