@weareoutman 2024-10-28
👍 Great work!

I got no idea why it doesn't work with yarn PnP, either. Though using node modules linker works, right? Maybe related to dependencies with canary versions.