@slorber 2024-10-28
I'll try again once v3.6 is out and we'll see

For now it's already nice to have a POC proving that this plugin works with the new bundler :)