@slorber 2024-11-04
Docusaurus v3.6 is out: https://docusaurus.io/blog/releases/3.6

PR ready to review

Unfortunately, the Yarn PnP error remains, and it only builds with node linker. Maybe @arcanis any idea what could cause this error? 

```
[ERROR] Client bundle compiled with errors therefore further build is impossible.
× Module not found: Can't resolve '/Users/sebastienlorber/Desktop/projects/docusaurus-search-local-easyops-cn/.yarn/__virtual__/@docusaurus-core-virtual-22c3351990/0/cache/@docusaurus-core-npm-3.6.0-794aa300d6-d024957005.zip/node_modules/@docusaurus/core/lib/client/clientEntry.js' in '/Users/sebastienlorber/Desktop/projects/docusaurus-search-local-easyops-cn/website'
```