@slorber 2025-01-17
Looks like Yarn PnP will be solved in Rspack v1.2 soon

https://github.com/web-infra-dev/rspack/pull/7639