@FoxxMD 2025-01-17
Hi :wave: not really an issue but I am a little confused...is this PR in order to make the package examples or build work with Faster (IE for the developer), or is it to provide a working plugin for the end-user who is using Faster? I ask because I can build my [docusaurus site with Faster, default settings, and this plugin](https://github.com/FoxxMD/multi-scrobbler/commit/d724bef4e4eeef831bcdfa193e6e57e0c7f3f5a2) without an issue.