@slorber 2025-01-17
This plugin is already compatible, I only turned Docusaurus Faster on the demo website as a proof that this is compatible.


See also my guide for plugins authors: https://github.com/facebook/docusaurus/issues/10572 

None of these changes are required for this plugin, as far as I remember.