@slorber 2025-01-27
Tried to upgrade this PR but still getting errors when using Rspack + Yarn PnP, while it works fine with Webpack + Yarn PnP

Errors such as:

> × Cannot find module '@mdx-js/react' for matched aliased key '@mdx-js/react'

> × Cannot find module 'react' for matched aliased key 'react'

> × Module not found: Can't resolve './src/standalone/' in '/Users/sebastienlorber/Desktop/projects/docusaurus-search-local-easyops-cn/.yarn/cache/@easyops-cn-autocomplete.js-npm-0.38.1-2375cf670a-40a053d7fb.zip/node_modules/@easyops-cn/autocomplete.js'
>    ╭─[1:0]
>  1 │ 'use strict';
>    · ▲
>  2 │ 
>  3 │ module.exports = require('./src/standalone/');
>    ╰────

> × Module not found: Can't resolve './BaseUrlIssueBanner' in '/Users/sebastienlorber/Desktop/projects/docusaurus-search-local-easyops-cn/.yarn/__virtual__/@docusaurus-core-virtual-8ba776b5dc/0/cache/@docusaurus-core-npm-3.7.0-64e9676bfa-df07928282.zip/node_modules/@docusaurus/core/lib/client'
>     ╭─[18:31]
>  16 │ import { DocusaurusContextProvider } from './docusaurusContext';
>  17 │ import PendingNavigation from './PendingNavigation';
>  18 │ import BaseUrlIssueBanner from './BaseUrlIssueBanner';
>     ·                                ──────────────────────
>  19 │ import SiteMetadataDefaults from './SiteMetadataDefaults';
>  20 │ // TODO, quick fix for CSS insertion order
>     ╰────
> 

> × Module not found: Can't resolve './components/ThemedComponent' in '/Users/sebastienlorber/Desktop/projects/docusaurus-search-local-easyops-cn/.yarn/__virtual__/@docusaurus-theme-common-virtual-67b7f3254f/0/cache/@docusaurus-theme-common-npm-3.7.0-5d72768bc3-9b27a2c74f.zip/node_modules/@docusaurus/theme-common/lib'
>     ╭─[49:43]
>  47 │  * APIs to document
>  48 │  */ export { useThemeConfig } from './utils/useThemeConfig';
>  49 │ export { default as ThemedComponent } from './components/ThemedComponent';
>     ·                                            ──────────────────────────────
>  50 │ export { createStorageSlot, useStorageSlot, listStorageKeys } from './utils/storageUtils';
>  51 │ export { usePluralForm } from './utils/usePluralForm';
>     ╰────


Not sure how to troubleshoot these problems but will have to investigate to know if the issues affect only this site/monorepo, or Docusaurus + Rspack in general.

cc @hardfist